* Fixed: A Number of Compilation Warnings in SDK.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4030 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/editarrayfiledlg.cpp
@@ -67,7 +67,7 @@ void EditArrayFileDlg::EndModal(int retCode)
 	{
 		wxListBox* list = XRCCTRL(*this, "lstItems", wxListBox);
 		m_Array.Clear();
-		for (int i = 0; i < list->GetCount(); ++i)
+		for (int i = 0; i < (int)list->GetCount(); ++i)
 		{
 			m_Array.Add(list->GetString(i));
 		}
