+++ src/sdk/editarraystringdlg.cpp
@@ -57,7 +57,7 @@ void EditArrayStringDlg::EndModal(int retCode)
 	{
 		wxListBox* list = XRCCTRL(*this, "lstItems", wxListBox);
 		m_Array.Clear();
-		for (int i = 0; i < list->GetCount(); ++i)
+		for (int i = 0; i < (int)list->GetCount(); ++i)
 		{
 			m_Array.Add(list->GetString(i));
 		}
