+++ src/sdk/multiselectdlg.cpp
@@ -79,7 +79,7 @@ void MultiSelectDlg::UpdateStatus()
 {
 	int count = 0;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         if (lst->IsChecked(i))
             ++count;
@@ -93,7 +93,7 @@ wxArrayString MultiSelectDlg::GetSelectedStrings() const
 {
     wxArrayString ret;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         if (lst->IsChecked(i))
             ret.Add(lst->GetString(i));
@@ -105,7 +105,7 @@ wxArrayInt MultiSelectDlg::GetSelectedIndices() const
 {
     wxArrayInt ret;
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         if (lst->IsChecked(i))
             ret.Add(i);
@@ -119,7 +119,7 @@ void MultiSelectDlg::SelectWildCard(const wxString& wild, bool select, bool clea
         return;
 	wxArrayString wilds = GetArrayFromString(wild, _T(";"));
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
 		if (clearOld || !lst->IsChecked(i))
 		{
@@ -167,7 +167,7 @@ void MultiSelectDlg::OnItemToggle(wxCommandEvent& /*event*/)
 void MultiSelectDlg::OnToggle(wxCommandEvent& /*event*/)
 {
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         lst->Check(i, !lst->IsChecked(i));
 	}
@@ -177,7 +177,7 @@ void MultiSelectDlg::OnToggle(wxCommandEvent& /*event*/)
 void MultiSelectDlg::OnSelectAll(wxCommandEvent& /*event*/)
 {
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         lst->Check(i, true);
 	}
@@ -187,7 +187,7 @@ void MultiSelectDlg::OnSelectAll(wxCommandEvent& /*event*/)
 void MultiSelectDlg::OnDeselectAll(wxCommandEvent& /*event*/)
 {
     wxCheckListBox* lst = XRCCTRL(*this, "lstItems", wxCheckListBox);
-	for (int i = 0; i < lst->GetCount(); ++i)
+	for (int i = 0; i < (int)lst->GetCount(); ++i)
 	{
         lst->Check(i, false);
 	}
